pyproject.toml
[project]
name = "dslc7"
version = "0.1.0"
description = "対話システムライブコンペティション7"
authors = [
    { name = "yuta0306", email = "yuta.20010306@gmail.com" },
    { name = "Taka008" },
    { name = "snjstudent" },
    { name = "p1n0k0" },
]
dependencies = [
    "opencv-python>=4.10",
    "py-feat>=0.6.2",
    "pika>=1.3.2",
    "azure-cognitiveservices-speech>=1.40.0",
    "mecab-python3>=1.0.9",
    "PyAudio>=0.2.14",
    "PyYAML>=6.0.2",
    "numpy<2",
    "openai<1.0.0",
    "google-cloud-speech>=2.27.0",
    "librosa>=0.10.2",
    "soundfile>=0.12.1",
    "pysptk>=1.0.1",
    "torch==2.0.1",
    "torchvision==0.15.2",
    "unidic-lite>=1.0.8",
    "scipy==1.10.0",
    "einops",
    "torchaudio==2.0.2",
    "torchsummary",
    "omegaconf",
    "torchmetrics",
    "websockets>=13.1",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dslc7"]
